requirements.txt
pytest==8.3.4
selenium~=4.29.0
allure-pytest==2.13.5
allure-python-commons~=2.13.5